ESC-430 factor out lead only test support
